requirements-dev.txt
boto3-stubs[sqs,s3]
coverage-badge
mypy
pytest
pytest-cov
pytest-mock
ruff